Proper "game-add" implementation
Took a while to work out but essentially, the `maintenance.py` methods will be for resetting Ryu numbers in the event that they aren't inserted properly.
It works by starting with Ryu and all the games he appears in, removing and reinserting relations as necessary. Unfortunately for the moment, it also backtracks and re-inserts relations that have been reset previously, but I think that can be fixed by keeping track of the Ryu number at the moment and only removing relations with a Ryu number >= the current. We'll see if we can get that working soon!