Proper path output and functionality
The function implementation I've been dreading to put in turned out to be not so bad!
An hour or two of debugging later we get a function that will give you the direct path from one character, through a given amount of games, all the way to Ryu!
It works sort of like a linked list, but with alternating objects (because Python is just a beast like that).
Oh and also more data. Can't have too much of that!