'Add to game' command implemented
So a few things happened in this commit, but yeah the main thing is this new command.
Essentially, let's say a game exists in the DB, but new DLC is announced that adds another character (or you missed one by accident). Now you can add them without manually opening that text file!
In addition, in that command and the 'Insert game' command in the frontend, I streamlined the character insertion process in a separate function.
Oh and as always, M O R E   D A T A ! ! ! ! !