Implement 'update game' method
Implementation and proper functionality of the "update game" method, which includes:
- Update the title of a game
- Update the release date of a game
Updates occur directly to the database (and related local files) so effects are noticed instantly.
In addition, all methods have been further abstracted for updateData().updateGame().
Also forgot to mention in last commit, but some methods (validDate() and removeIllegalChars()) added for ease.