Implement 'update character' features
Now you can update a character's name in the database.
If the name you update it with is another character already in the db, it merges those two records.
Changes are also reflected locally.
Also, for the record, I Do Not like Python's file read/write, so I wrote a method to do it for me!