Add ability to see characters in game
Now after you query a game, you have the option to see the characters in that game.
It will bring up a resultViewer (that you cannot select characters from).
In addition, minor bug fix for the "..." part of the pages on the resultViewer.
And, yeah, more data (and data updates!)