Centralize query sanitization
Previously, the onus is on the caller of a query to remember to sanitize.
This lead to lower cohesion, so I moved the sanitization to the queries.
Now, functions can be called and the caller can be satisfied in knowing their data will be satisfied properly.
Also, did a little restructuring, that's why all these game files are being updated lol